requirements.txt
flask
flask-restplus
flask-login
passlib
flask-jwt-extended
enum34
sqlalchemy
simplejson
psycopg2-binary
flask-cors
requests
geos
geojson
pyproj
shapely
cx_Oracle
#bjoern
#MySQL-python
mysqlclient
future